services/behavioral-nudge-engine/requirements.txt
# Core
fastapi==0.95.0
uvicorn[standard]==0.21.1
python-dotenv==1.0.0
pydantic==1.10.7

# Database
sqlalchemy==2.0.9
alembic==1.10.3
psycopg2-binary==2.9.6
redis==4.5.5

# Utils
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
requests==2.28.2

# AI/ML
numpy==1.24.3
pandas==2.0.1
scikit-learn==1.2.2

# Testing
pytest==7.3.1
pytest-cov==4.0.0
httpx==0.23.3

# Code Quality
black==23.3.0
isort==5.12.0
mypy==1.2.0
flake8==6.0.0

# Monitoring
prometheus-client==0.16.0
opentelemetry-api==1.16.0
opentelemetry-sdk==1.16.0
opentelemetry-instrumentation-fastapi==0.37b0
